setFirmID & checkFirmCredentials moved to firmHandler
